requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.2
attrs==25.3.0
azure-ai-agents==1.0.2
azure-ai-evaluation==1.9.0
azure-ai-projects==1.0.0b12
azure-core==1.35.0
azure-identity==1.23.1
azure-monitor-opentelemetry-exporter==1.0.0b40
azure-storage-blob==12.26.0
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
distro==1.9.0
docstring_parser==0.16
fastapi==0.116.1
filelock==3.18.0
filetype==1.2.0
fixedint==0.1.6
Flask==3.1.1
flask-cors==6.0.1
flask-restx==1.3.0
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.70.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
isodate==0.7.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.1
jsonschema-specifications==2025.4.1
keyring==24.3.1
MarkupSafe==3.0.2
marshmallow==3.26.1
more-itertools==10.7.0
msal==1.32.3
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.6.3
nltk==3.9.1
numpy==2.3.1
oauthlib==3.3.1
openai==1.97.0
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-http==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
packaging==25.0
pandas==2.3.1
pillow==11.0.0
promptflow-core==1.18.1
promptflow-devkit==1.18.1
promptflow-tracing==1.18.1
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydash==7.0.7
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.26.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
SecretStorage==3.3.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
strictyaml==1.7.3
tabulate==0.9.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
waitress==3.0.2
Werkzeug==3.1.3
yarl==1.20.1
zipp==3.23.0
